[bazel] Allow opentitan_functest to create example instead of tests
The purpose of an example is to run a binary like an hello world that
does not necessarily have a nothing of pass/fail. In practice, this
means that we want to generate a sh_binary instead of a sh_test and
not generate a test suit.

Since bitstreams and maybe artefacts are marked as test only, I opted
to make the sh_binary produced by opentitan_functest also test-only
so that bazel does not complain about non-tests targets depend
on test targets.

Signed-off-by: Amaury Pouly <amaury.pouly@lowrisc.org>